refactor(native): isolate sidecar archive tests
